+++ js/constans.js
@@ -1,0 +1,51 @@
+const POSTS_COUNT = 25;
+
+const NAMES = [
+  'Милан',
+  'Тарас',
+  'Ира',
+  'Кристина',
+  'Саша',
+  'Юлия',
+  'Люба',
+  'Миша',
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const DESCRIPTIONS = [
+  'лагерь для отдыха',
+  'добро пожаловать на пляж',
+  'а вот и пляж',
+  'мощный фото_аппарат',
+  'пошли плавать',
+  'люблю матовую краску',
+  'минималистичный десерт',
+  'виноградный компот',
+  'подвезли свеженькое пивко',
+  'также хочу такую полочку для обуви',
+  'вход в рай',
+  'а "мерен" все таки лучше',
+  'наконец, полезная еда',
+  'и еще немного полезной пищи',
+  'мои скороходы',
+  'пора домой',
+  'местный КВН',
+  'легенда',
+  'чего только не придумают эти китайцы',
+  'круглый год на них смотрел бы',
+  'угадайте что это за блюдо',
+  'такой закат солнца больше нигде не увидишь',
+  'деликатес)',
+  'сегодня ночью будут танцы',
+  'советую всем попробовать сафари',
+];
+
+export {POSTS_COUNT, NAMES, COMMENTS, DESCRIPTIONS}
